+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 * Various code quality fixes using FindBugs & Coverity
+* Make audio stream playing work (example: SndPlayer.java)
 
 ## v1.5.0 (2016-09-02)